@@ -1,6 +1,6 @@
 module github.com/roadrunner-server/kv/v2
 
-go 1.18
+go 1.19
 
 require (
 	github.com/roadrunner-server/api/v2 v2.20.0
